requirement.txt
python-telegram-bot
gpiozero
